epoll: cache alignment for per-thread structure
This probably won't make a huge difference in Ruby, but perhaps
one day the unnecessary dirtying of cache lines will affect
performance (and we'll be ready when that day comes).

While we're at it, remove usage of pthread* functions for
thread-local variables.  The __thread construct from GCC (and
also implemented by clang) is much easier-to-use than the
pthread_*specific API.