license + gem packaging cleanups
require_paths and date are automatically set in modern RubyGems
versions.  Since we only use modern RubyGems versions, the
licenses= accessor is also activated.

Since the FSF may change mailing addresses again in the future,
prefer the web address to point to licenses.  This change is
acceptable for GNU projects, so it should be for us, too:
http://sourceware.org/bugzilla/show_bug.cgi?id=13673