epoll: cleanup GVL-release code for Ruby 2.0.0
rb_thread_blocking_region is deprecated in Ruby 2.0, but
rb_thread_io_blocking region is not (and superior for I/O).  So we will
favor rb_thread_io_blocking_region for now.

While we're at it, reimplement green-thread-safe (Ruby 1.8) epoll_wait
in Ruby instead of C.  The extra #ifdefs for 1.8.7 were more prone to
bitrot and Ruby code should be easier to follow for Rubyists who care
about 1.8.